// src/types/constants.ts
export enum APPLICATION_COMMAND_OPTION_TYPES {
    SUB_COMMAND = 1,
    SUB_COMMAND_GROUP = 2,
    STRING = 3,
    INTEGER = 4,
    BOOLEAN = 5,
    USER = 6,
    CHANNEL = 7,
    ROLE = 8,
    MENTIONABLE = 9,
    NUMBER = 10,
}

export enum Colors {
    DEFAULT = 0x000000,
    WHITE = 0xFFFFFF,
    AQUA = 0x1ABC9C,
    GREEN = 0x2ECC71,
    BLUE = 0x3498DB,
    YELLOW = 0xFFFF00,
    PURPLE = 0x9B59B6,
    LUMINOUS_VIVID_PINK = 0xE91E63,
    GOLD = 0xF1C40F,
    ORANGE = 0xE67E22,
    RED = 0xE74C3C,
    GREY = 0x95A5A6,
    NAVY = 0x34495E,
    DARK_AQUA = 0x11806A,
    DARK_GREEN = 0x1F8B4C,
    DARK_BLUE = 0x206694,
    DARK_PURPLE = 0x71368A,
    DARK_VIVID_PINK = 0xAD1457,
    DARK_GOLD = 0xC27C0E,
    DARK_ORANGE = 0xA84300,
    DARK_RED = 0x992D22,
    DARK_GREY = 0x979C9F,
    DARKER_GREY = 0x7F8C8D,
    LIGHT_GREY = 0xBCC0C0,
    DARK_NAVY = 0x2C3E50,
    BLURPLE = 0x7289DA,
    GREYPLE = 0x99AAB5,
    DARK_BUT_NOT_BLACK = 0x2C2F33,
    NOT_QUITE_BLACK = 0x23272A
}

export enum PERMISSIONS {
    CREATE_INSTANT_INVITE = 0x00000001,
    KICK_MEMBERS = 0x00000002,
    BAN_MEMBERS = 0x00000004,
    ADMINISTRATOR = 0x00000008,
    MANAGE_CHANNELS = 0x00000010,
    MANAGE_GUILD = 0x00000020,
    ADD_REACTIONS = 0x00000040,
    VIEW_AUDIT_LOG = 0x00000080,
    PRIORITY_SPEAKER = 0x00000100,
    STREAM = 0x00000200,
    VIEW_CHANNEL = 0x00000400,
    SEND_MESSAGES = 0x00000800,
    SEND_TTS_MESSAGES = 0x00001000,
    MANAGE_MESSAGES = 0x00002000,
    EMBED_LINKS = 0x00004000,
    ATTACH_FILES = 0x00008000,
    READ_MESSAGE_HISTORY = 0x00010000,
    MENTION_EVERYONE = 0x00020000,
    USE_EXTERNAL_EMOJIS = 0x00040000,
    VIEW_GUILD_INSIGHTS = 0x00080000,
    CONNECT = 0x00100000,
    SPEAK = 0x00200000,
    MUTE_MEMBERS = 0x00400000,
    DEAFEN_MEMBERS = 0x00800000,
    MOVE_MEMBERS = 0x01000000,
    USE_VAD = 0x02000000,
    CHANGE_NICKNAME = 0x04000000,
    MANAGE_NICKNAMES = 0x08000000,
    MANAGE_ROLES = 0x10000000,
    MANAGE_WEBHOOKS = 0x20000000,
    MANAGE_EMOJIS = 0x40000000,
    USE_SLASH_COMMANDS = 0x80000000,
    REQUEST_TO_SPEAK = 0x100000000,
    MANAGE_THREADS = 0x400000000,
    USE_PUBLIC_THREADS = 0x800000000,
}